--- packages/esm-patient-queues-app/src/active-visits/active-visits-patients-reception/active-visits-reception.resource.ts
@@ -1,48 +1,0 @@
-import { formatDate, openmrsFetch, parseDate, restBaseUrl } from '@openmrs/esm-framework';
-import dayjs from 'dayjs';
-import useSWR from 'swr';
-import { PatientQueue } from '../../types/patient-queues';
-
-export function usePatientQueuesList(currentQueueRoomLocationUuid: string) {
-  const apiUrl = `${restBaseUrl}/patientqueue?v=full&status=pending&parentLocation=${currentQueueRoomLocationUuid}`;
-  const { data, error, isLoading, isValidating, mutate } = useSWR<{ data: { results: Array<PatientQueue> } }, Error>(
-    apiUrl,
-    openmrsFetch,
-  );
-
-  const mapppedQueues = data?.data?.results.map((queue: PatientQueue) => {
-    return {
-      ...queue,
-      id: queue.uuid,
-      name: queue.patient?.person.display,
-      patientUuid: queue.patient?.uuid,
-      priorityComment: queue.priorityComment,
-      priority: queue.priorityComment === 'Urgent' ? 'Priority' : queue.priorityComment,
-      waitTime: queue.dateCreated ? `${dayjs().diff(dayjs(queue.dateCreated), 'minutes')}` : '--',
-      status: queue.status,
-      patientAge: queue.patient?.person?.age,
-      patientSex: queue.patient?.person?.gender === 'M' ? 'MALE' : 'FEMALE',
-      patientDob: queue.patient?.person?.birthdate
-        ? formatDate(parseDate(queue.patient.person.birthdate), { time: false })
-        : '--',
-      identifiers: queue.patient?.identifiers,
-      locationFrom: queue.locationFrom?.uuid,
-      locationTo: queue.locationTo?.uuid,
-      queueRoom: queue.locationTo?.display,
-      visitNumber: queue.visitNumber,
-      dateCreated: queue.dateCreated,
-      creatorUuid: queue.creator?.uuid,
-      creatorUsername: queue.creator?.username,
-      creatorDisplay: queue.creator?.display,
-    };
-  });
-
-  return {
-    patientQueueEntries: mapppedQueues || [],
-    patientQueueCount: mapppedQueues?.length,
-    isLoading,
-    isError: error,
-    isValidating,
-    mutate,
-  };
-}
