--- packages/esm-patient-queues-app/src/active-visits/active-visits-patients-reception/active-visits-reception.scss
@@ -1,247 +1,0 @@
-@use '@carbon/styles/scss/type';
-@use '@carbon/styles/scss/spacing';
-@import '~@openmrs/esm-styleguide/src/vars';
-
-.container {
-  margin-bottom: 30px;
-}
-
-.headerContainer {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 1rem;
-  background-color: $ui-background;
-}
-
-.headerButtons {
-  display: flex;
-  flex-flow: column;
-}
-
-.heading {
-  font-size: 20px;
-  font-weight: bold;
-}
-
-.filterContainer {
-  :global(.cds--dropdown__wrapper--inline) {
-    gap: 0;
-  }
-
-  :global(.cds--list-box__menu-icon) {
-    height: 1rem;
-  }
-
-  :global(.cds--dropdown--inline) :global(.cds--list-box__field) {
-    min-width: 12rem;
-  }
-}
-
-.tooltip :global(.cds--tooltip__trigger.cds--tooltip__trigger--definition) {
-  border-bottom: none;
-}
-
-.tag {
-  margin: 0.25rem 0;
-}
-
-.priorityTag {
-  @extend .tag;
-  @include type.type-style('label-01');
-  color: #943d00;
-  background-color: #ffc9a3;
-}
-
-
-
-.search {
-  max-width: 16rem;
-
-  input {
-    background-color: $ui-02 !important;
-  }
-}
-
-.tableContainer {
-  border: 1px solid $ui-03;
-  margin: 0 spacing.$spacing-05;
-  padding: 0;
-
-  a {
-    text-decoration: none;
-  }
-
-  th {
-    color: $text-02;
-  }
-
-  :global(.cds--data-table) {
-    background-color: $ui-03;
-  }
-
-  .toolbarContent {
-    height: spacing.$spacing-07;
-    margin-bottom: spacing.$spacing-02;
-  }
-
-  :global(.cds--toolbar-search-container-active.cds--search) {
-    margin-right: 1rem;
-  }
-}
-
-.activeVisitsTable tr:last-of-type {
-  td {
-    border-bottom: none;
-  }
-}
-
-.expandedActiveVisitRow {
-  :global(.cds--tab-content) {
-    padding: 0.5rem 0;
-  }
-
-  td {
-    padding: 0.5rem;
-
-    >div {
-      max-height: max-content !important;
-      background-color: $ui-02;
-    }
-  }
-
-  th[colspan] td[colspan]>div:first-child {
-    padding: 0 1rem;
-  }
-}
-
-.hiddenRow {
-  display: none;
-}
-
-.content {
-  @include type.type-style('heading-compact-02');
-  color: $text-02;
-  margin-bottom: 0.5rem;
-}
-
-.helper {
-  @include type.type-style('body-compact-01');
-  color: $text-02;
-}
-
-.separator {
-  @include type.type-style('body-compact-02');
-  color: $text-02;
-  width: 80%;
-  margin: 1.5rem auto;
-  overflow: hidden;
-  text-align: center;
-
-  &::before,
-  &::after {
-    background-color: $text-03;
-    content: '';
-    display: inline-block;
-    height: 1px;
-    position: relative;
-    vertical-align: middle;
-    width: 50%;
-  }
-
-  &::before {
-    right: 0.5rem;
-    margin-left: -50%;
-  }
-
-  &::after {
-    left: 0.5rem;
-    margin-right: -50%;
-  }
-}
-
-.tileContainer {
-  background-color: $ui-02;
-  border-top: 1px solid $ui-03;
-  padding: 5rem 0;
-}
-
-.tile {
-  margin: auto;
-  width: fit-content;
-}
-
-.tileContent {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.menuItem {
-  max-width: none;
-}
-
-.desktopHeading {
-  h4 {
-    @include type.type-style('heading-compact-02');
-    color: $text-02;
-  }
-}
-
-.tabletHeading {
-  h4 {
-    @include type.type-style('heading-03');
-    color: $text-02;
-  }
-}
-
-.desktopHeading,
-.tabletHeading {
-  text-align: left;
-  text-transform: capitalize;
-  margin-bottom: spacing.$spacing-05;
-
-  h4:after {
-    content: '';
-    display: block;
-    width: 2rem;
-    padding-top: 3px;
-    border-bottom: 0.375rem solid var(--brand-03);
-  }
-}
-
-.statusContainer {
-  display: flex;
-  align-items: center;
-
-  svg {
-    margin-right: 0.5rem;
-  }
-}
-
-.visitType {
-  @include type.type-style('heading-compact-02');
-}
-
-.headerBtnContainer {
-  background-color: $ui-background;
-  padding: spacing.$spacing-05;
-  text-align: right;
-}
-
-.addPatientToListBtn {
-  margin-left: spacing.$spacing-05;
-  height: spacing.$spacing-09;
-}
-
-.editIcon {
-  color: $interactive-01;
-  margin-top: 0.5rem;
-  cursor: pointer;
-}
-
-.toolbarDiv {
-  display: flex;
-  justify-content: space-between;
-  margin: 2rem;
-}
