+++ packages/esm-patient-queues-app/src/queue-clinical-room-home.component.tsx
@@ -3,7 +3,7 @@ import React, { useState } from 'react';
 import PatientQueueHeader from './components/patient-queue-header/patient-queue-header.component';
 import { useTranslation } from 'react-i18next';
 import { QueueStatus } from './utils/utils';
-import ActiveClinicalVisitsTable from './active-visits/active-visits-patients-clinical/active-visits-clinical-table.component';
+import ActiveClinicalVisitsTable from './active-visits/queue-patients-clinical/queue-clinical-table.component';
 import styles from './active-visits/active-visits-table.scss';
 import { Tabs, TabPanel, TabList, Tab, TabPanels } from '@carbon/react';
